remaining changes for move extra logic from home to guards
