+++ website/src/components/EncryptionKeyInput.vue
@@ -2,20 +2,18 @@
 import { ref } from "vue";
 import LogoButton from "@/components/Logo/LogoButton.vue";
 import { useUserInfoStore } from "@/piniaStore/userInfo";
-import { onMounted } from "vue";
 
 const emits = defineEmits<{
   (e: "success", encryptionKey: string): void;
 }>();
 
-const { validateEncryptionKey, loadUserInfoIfRequred } = useUserInfoStore();
+const { validateEncryptionKey } = useUserInfoStore();
 
 const encryptionKey = ref("");
 const warningMessage = ref("");
 const errorMessage = ref("");
 
 const isFormValid = ref(false);
-const loading = ref(false);
 
 function submitHandler() {
   if (isFormValid.value == false || isFormValid.value == null) {
@@ -27,21 +25,6 @@ function submitHandler() {
   }
   errorMessage.value = "wrong encryption key";
 }
-onMounted(() => {
-  loading.value = true;
-  loadUserInfoIfRequred()
-    .then((ok) => {
-      if (!ok) {
-        console.error("something went wrong while loading user info");
-        errorMessage.value = "something went wrong while loading user info";
-      }
-      loading.value = false;
-    })
-    .catch((err) => {
-      console.error(err);
-      loading.value = false;
-    });
-});
 </script>
 <template>
   <v-sheet
@@ -70,12 +53,14 @@ onMounted(() => {
                 label="encryption key"
                 type="password"
                 required
-                :rules="[(value: string) => {
-                  if (value.length > 0) {
-                    return true
-                  }
-                  return 'required'
-                }]"
+                :rules="[
+                  (value: string) => {
+                    if (value.length > 0) {
+                      return true;
+                    }
+                    return 'required';
+                  },
+                ]"
                 autocomplete="on"
                 prepend-inner-icon="mdi-lock-outline"
               ></v-text-field>
