+++ website/src/js/serviceWorker/registeration.ts
@@ -1,5 +1,5 @@
-// import decryptWorker from "@/worker/dist/bundle.js?url";
-import decryptWorker from "@/worker/decrypt?url";
+import decryptWorker from "@/worker/dist/bundle.js?url";
+// import decryptWorker from "@/worker/decrypt?url";
 
 export function updateOrRegisterServiceWorker(): Promise<ServiceWorker> {
   return new Promise<ServiceWorker>((resolve, reject) => {
